refactor: user-list,profileページをfeaturesディレクトリ化
