+++ frontend/src/features/users/index.ts
@@ -1,0 +1,2 @@
+export * from './routes/UserList';
+export * from './routes/Profile';
