+++ frontend/src/features/users/routes/UserList.tsx
@@ -7,8 +7,8 @@ import {
   Thead,
   TableContainer,
 } from '@chakra-ui/react';
-import { UserTableData } from 'components/organisms/user/UserTableData';
-import { useAllUsers } from '../../hooks/useAllUsers';
+import { UserTableData } from 'features/users/components/UserTableData';
+import { useAllUsers } from '../api/useAllUsers';
 
 export const UserList: FC = memo(() => {
   const { getUsers, users } = useAllUsers();
@@ -26,8 +26,8 @@ export const UserList: FC = memo(() => {
             {users.map((users) => (
               <UserTableData
                 key={users.id}
-                login={users.name}
                 imageUrl="https://source.unsplash.com/random"
+                login={users.name}
               />
             ))}
           </Tbody>
