--- frontend/src/types/api/accessToken.ts
@@ -1,3 +1,0 @@
-export type AccessToken = {
-  accessToken: string;
-};
